requirements.txt
# Core dependencies
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# File processing
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0
pytesseract>=0.3.10  # Optional: For OCR on images (requires Tesseract OCR installed)

# Audio/Video processing
yt-dlp>=2023.10.7
pydub>=0.25.1
speechrecognition>=3.10.0  # For audio transcription (requires internet or offline engine)

# Database (sqlite3 is built-in to Python)

# Utilities
requests>=2.31.0
numpy>=1.24.0

